Use sudo instead of su.
See https://github.com/dotcloud/docker/issues/2972:

The issue describes a problem where an `scp` failing to reach the remote host
make the whole container exit. Using sudo instead of su is a possible
workaround. `scp` can be used to archive WAL segment files in the
archive_command in the PostgreSQL configuration file so we want to make sure it
can work.