Make it possible to create /data with a script.
This also fix the use of `postgres -D`. The official documentation says in one
page it is use to provide the path to the data directory. But in another page
it says it provides the path to the directory containing the configuration.
Anyway postgres.conf specifies the data directory so we use `-c data_directory`
instead.